User Auth created, File upload S3 connected working

+++ src/controllers/pizzas_controller.py
@@ -1,22 +1,25 @@
 from models.Pizza import Pizza
+from models.User import User
 from main import db
 from schemas.PizzaSchema import pizza_schema, pizzas_schema
 from flask import Blueprint, request, jsonify, abort
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from services.auth_service import verify_user
+from sqlalchemy.orm import joinedload
+
 pizzas = Blueprint('pizzas', __name__, url_prefix="/pizzas")
-
-
-#Retrieve all pizzas
 
 @pizzas.route("/", methods=["GET"])
 def pizza_index():
-    pizzas = Pizza.query.all()
+    #Retrieve all pizzas
+    pizzas = Pizza.query.options(joinedload("user")).all()
     return jsonify(pizzas_schema.dump(pizzas))
 
-
-#Create a new pizza
-
 @pizzas.route("/", methods=["POST"])
-def pizza_create():
+@jwt_required
+@verify_user
+def pizza_create(user=None):
+    #Create a new pizza
     pizza_fields = pizza_schema.load(request.json)
 
     new_pizza = Pizza()
@@ -24,41 +27,47 @@ def pizza_create():
     new_pizza.description = pizza_fields["description"]
     new_pizza.price = pizza_fields["price"]
     new_pizza.location = pizza_fields["location"]
-    new_pizza.pizza_image = pizza_fields["pizza_image"]
-       
-    db.session.add(new_pizza)
+
+    user.pizzas.append(new_pizza)
+
     db.session.commit()
     
-    return (jsonify(pizza_schema.dump(new_pizza)), 200)
+    return jsonify(pizza_schema.dump(new_pizza))
 
-
-#Return a single pizza
-
-@pizzas.route("/<int:pizza_id>", methods=["GET"])
-def pizza_show(pizza_id):
-    pizza = Pizza.query.get(pizza_id)
+@pizzas.route("/<int:id>", methods=["GET"])
+def pizza_show(id):
+    #Return a single pizza
+    pizza = Pizza.query.get(id)
     return jsonify(pizza_schema.dump(pizza))
 
-
-#Update a pizza
-
-@pizzas.route("/<int:pizza_id>", methods=["PUT", "PATCH"])
-def pizza_update(pizza_id):
-    
-    pizzas = Pizza.query.filter_by(pizza_id=pizza_id)
+@pizzas.route("/<int:id>", methods=["PUT", "PATCH"])
+@jwt_required
+@verify_user
+def pizza_update(id, user=None):
+    #Update a pizza
     pizza_fields = pizza_schema.load(request.json)
+
+    pizzas = Pizza.query.filter_by(pizza_id=id, user_id=user.user_id)
+
+    if pizzas.count() != 1:
+        return abort(401,  description="Unauthorized to update this pizza")    
+    
     pizzas.update(pizza_fields)
     db.session.commit()
 
     return jsonify(pizza_schema.dump(pizzas[0]))
-    
 
-#Delete a pizza
+@pizzas.route("/<int:id>", methods=["DELETE"])
+@jwt_required
+@verify_user
+def pizza_delete(id, user=None):
+    #Delete a pizza
 
-@pizzas.route("/<int:pizza_id>", methods=["DELETE"])
-def pizza_delete(pizza_id):
+    pizza = Pizza.query.filter_by(pizza_id=id, user_id=user.user_id).first()
+
+    if not pizza:
+        return abort(400)
     
-    pizza = Pizza.query.get(pizza_id)
     db.session.delete(pizza)
     db.session.commit()
